+++ spec/shared_behaviours/iterative_functions.rb
